pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dj-rest-admin"
version = "1.0.1"
description = "A package to generate CRUD endpoints for registered models with the Django-REST Framework."
readme = "README.md"
authors = [{ name = "Dario Fragas", email = "dariofg98@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
     "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.8",
        'Intended Audience :: Developers',  # Define that your audience are developers
        'Topic :: Software Development :: Build Tools'
]
keywords=['django', 'python', 'django-rest-framework', "admin", "api"]
dependencies = [
    "django >= 3.2",
    "djangorestframework >= 3.13"
]
requires-python = ">=3.8"


[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest", "coverage"]

[project.urls]
Homepage = "https://github.com/dfg-98/dj-rest-admin"


[tool.setuptools.packages.find]
include = ["dj_rest_admin*"]  

[tool.flake8]
max-line-lenght=80
